Refactoring, fixed some stuff.
