+++ plc4j/core/pom.xml
@@ -61,6 +61,11 @@
       <artifactId>netty-transport</artifactId>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-lang3</artifactId>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 
 </project>
